requirements.txt
Customtkinter
Pillow
numpy